# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Fingarden API Docs",
    "version": "1.0.0",
    "title": "Fingarden"
  },
  "basePath": "/",
  "paths": {
...
    "/products/{productId}": {
      "get": {
        "tags": ["product"],
        "summary": "Find product by id",
        "operationId": "findProductById",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "put": {
        "tags": ["product"],
        "summary": "Update the product",
        "description": "",
        "operationId": "updateProduct",
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin-token"
          },
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/ProductDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful updated",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/ProductWithoutJoins"
            }
          },
          "400": {
            "description": "Invalid Product request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "delete": {
        "tags": ["product"],
        "summary": "Delete the product",
        "description": "",
        "operationId": "deleteProduct",
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin-token"
          },
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful deleted",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "desc": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string",
                  "readOnly": true
                },
                "updatedAt": {
                  "type": "string",
                  "readOnly": true
                },
                "url": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
...
    "/orderProducts": {
      "get": {
        "tags": ["orderProduct"],
        "summary": "Get array of orderProducts",
        "description": "",
        "operationId": "getOrderProducts",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin-token or user-token"
          },
          {
            "name": "productId",
            "in": "query",
            "description": "Filtering orderProducts by productId",
            "required": false,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "minQty",
            "in": "query",
            "description": "Filtering orderProducts by qty greater then minQty",
            "required": false,
            "type": "integer"
          },
          {
            "name": "maxQty",
            "in": "query",
            "description": "Filtering orderProducts by qty less then maxQty",
            "required": false,
            "type": "integer"
          },
          {
            "name": "minPrice",
            "in": "query",
            "description": "Filtering orderProducts by product price greater then minPrice",
            "required": false,
            "type": "integer"
          },
          {
            "name": "maxPrice",
            "in": "query",
            "description": "Filtering orderProducts by product price less then maxPrice",
            "required": false,
            "type": "integer"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sorting orderProducts by property",
            "default": "productId",
            "required": false,
            "type": "string"
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Ordering results. ASC | DESC",
            "default": "DESC",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limiting orderProducts in array",
            "default": 10,
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OrderProductResponse"
              }
            }
          }
        }
      }
    },
    "/orderProducts/{orderProductId}": {
      "get": {
        "tags": ["orderProduct"],
        "summary": "Find orderProduct by id",
        "operationId": "findOrderProductById",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "type": "string",
            "name": "authorization",
            "required": true,
            "in": "header",
            "description": "admin-token or user-token"
          },
          {
            "name": "orderProductId",
            "in": "path",
            "description": "ID of orderProduct to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/OrderProduct"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "OrderProduct not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/api/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "Brand": {
      "required": ["name", "url"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "showOnMain": {
          "type": "boolean",
          "default": false,
          "description": "value true can be post only with admin token"
        }
      },
      "xml": {
        "name": "Brand"
      }
    },
...
    "Category": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "parent": {
          "type": "object",
          "$ref": "#/definitions/Category"
        },
        "children": {
          "type": "object",
          "$ref": "#/definitions/Category"
        },
        "parameters": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Parameter"
          }
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "Color": {
      "required": ["name", "url", "code"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ProductWithoutJoins"
          },
          "readOnly": true
        },
        "url": {
          "type": "string"
        },
        "code": {
          "type": "string"
        }
      }
    },
...
    "Parameter": {
      "required": ["name"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Rating": {
      "properties": {
        "1": {
          "type": "number"
        },
        "2": {
          "type": "number"
        },
        "3": {
          "type": "number"
        },
        "4": {
          "type": "number"
        },
        "5": {
          "type": "number"
        },
        "avg": {
          "type": "number"
        }
      }
    },
    "ParameterProduct": {
      "properties": {
        "productId": {
          "type": "string"
        },
        "parameterId": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "parameter": {
          "$ref": "#/definitions/Parameter"
        }
      }
    },
    "ParameterProductsDTO": {
      "required": ["parameterId", "value"],
      "properties": {
        "parameterId": {
          "type": "number"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "Product": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "shortDesc": {
          "type": "string"
        },
        "keywords": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "category": {
          "type": "object",
          "$ref": "#/definitions/Category"
        },
        "brand": {
          "type": "object",
          "$ref": "#/definitions/Brand"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/TagWithoutJoins"
          }
        },
        "sizes": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/SizeWithoutJoins"
          }
        },
        "url": {
          "type": "string"
        },
        "rating": {
          "$ref": "#/definitions/Rating"
        },
        "reviews": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewWithoutJoins"
          }
        },
        "questions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Question"
          }
        },
        "parameterProducts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ParameterProduct"
          }
        },
        "productVariants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductVariant"
          }
        }
      }
    },
...
    "ProductWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "url": {
          "type": "string"
        }
      }
    },
    "ProductVariant": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "price": {
          "type": "integer",
          "minimum": 1
        },
        "oldPrice": {
          "type": "integer"
        },
        "wholeSalePrice": {
          "type": "integer"
        },
        "available": {
          "type": "boolean"
        },
        "images": {
          "type": "string"
        },
        "color": {
          "type": "object",
          "$ref": "#/definitions/Color"
        }
      }
    },
    "ProductDTO": {
      "required": ["name", "category", "brand", "url"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "shortDesc": {
          "type": "string"
        },
        "keywords": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "array of tags IDs"
        },
        "category": {
          "type": "string",
          "description": "Category ID"
        },
        "brand": {
          "type": "string",
          "description": "Brand ID"
        },
        "url": {
          "type": "string"
        },
        "parameterProducts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ParameterProductsDTO"
          }
        },
        "productVariants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductVariant"
          }
        }
      }
    },
...
    "TagWithoutJoins": {
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "SizeWithoutJoins": {
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "Basket": {
      "required": ["orderProducts"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "userId": {
          "type": "string"
        },
        "orderProducts": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/OrderProduct"
          }
        },
        "totalAmount": {
          "type": "number"
        },
        "checkout": {
          "type": "object",
          "$ref": "#/definitions/CheckoutWithoutJoins"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "updatedAt": {
          "type": "string",
          "readOnly": true
        },
        "status": {
          "type": "string",
          "default": "New"
        }
      }
    },
...
    "CheckoutWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "paymentId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        },
        "leaveNearDoor": {
          "type": "boolean"
        }
      }
    },
...
    "OrderProduct": {
      "required": ["product", "inBasket"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "product": {
          "type": "object",
          "$ref": "#/definitions/Product"
        },
        "productVariant": {
          "type": "object",
          "$ref": "#/definitions/ProductVariant"
        },
        "qty": {
          "type": "integer",
          "minimum": 1
        },
        "productSize": {
          "type": "string"
        },
        "productPrice": {
          "type": "integer",
          "readOnly": true
        },
        "inBasket": {
          "type": "object",
          "$ref": "#/definitions/Basket"
        }
      }
    },
    "OrderProductResponse": {
      "properties": {
        "rows": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderProductWithJoins"
          }
        },
        "length": {
          "type": "number"
        }
      }
    },
...
    "OrderProductWithJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "product": {
          "type": "object",
          "$ref": "#/definitions/Product"
        },
        "productVariant": {
          "type": "object",
          "$ref": "#/definitions/ProductVariant"
        },
        "qty": {
          "type": "integer",
          "minimum": 1
        },
        "productPrice": {
          "type": "integer",
          "readOnly": true
        },
        "inBasket": {
          "type": "object",
          "$ref": "#/definitions/Basket"
        }
      }
    },
...
    "Review": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "rating": {
          "type": "integer",
          "minimum": 1,
          "maximum": 10
        },
        "text": {
          "type": "string"
        },
        "images": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "showOnMain": {
          "type": "boolean",
          "default": false
        },
        "product": {
          "type": "object",
          "$ref": "#/definitions/Product"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "reactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewReaction"
          }
        }
      }
    },
    "Question": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "text": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "product": {
          "type": "object",
          "$ref": "#/definitions/Product"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/QuestionComment"
          }
        },
        "reactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReactionQuestion"
          }
        }
      }
    },
...
    "ReviewWithoutJoins": {
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "rating": {
          "type": "integer",
          "minimum": 1,
          "maximum": 10
        },
        "text": {
          "type": "string"
        },
        "images": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "showOnMain": {
          "type": "boolean",
          "default": false
        },
        "productId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "reactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewReaction"
          }
        }
      }
    },
...
    "ReviewReaction": {
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "reviewId": {
          "type": "string"
        },
        "reaction": {
          "type": "string",
          "default": "Like"
        }
      }
    },
    "ReactionQuestion": {
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "questionId": {
          "type": "string"
        },
        "reaction": {
          "type": "string",
          "default": "Like"
        }
      }
    },
    "User": {
      "required": ["firstName", "lastName", "email", "password"],
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "isVerified": {
          "type": "boolean",
          "default": false
        },
        "role": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "Comment": {
      "properties": {
        "id": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "review": {
          "type": "object",
          "$ref": "#/definitions/Review"
        },
        "text": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "reactions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/CommentReaction"
          }
        }
      }
    },
    "QuestionComment": {
      "properties": {
        "id": {
          "type": "string"
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/User"
        },
        "question": {
          "type": "object",
          "$ref": "#/definitions/Question"
        },
        "text": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "reactions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/CommentReaction"
          }
        }
      }
    },
    "CommentReaction": {
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "commentId": {
          "type": "string"
        },
        "reaction": {
          "type": "string",
          "default": "Like"
        }
      }
    },
...
  }
}